SPARK-32548 - Review comments are addressed
